requirements.txt
matplotlib
scikit-learn
pygame
playsound
gTTS
myo-sdk-win-0.9.0